compile in non-functional TermIOS Qore class on platforms that do not support it so that programs can possibly still parse and execute; the option constant can be used to print out a user-friendly message
git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/qore/trunk@4244 1b0b35fe-f740-4c1b-b404-a338b32e513b

+++ lib/thread.cpp
@@ -145,7 +145,7 @@ public:
 #if (defined _WIN32 || defined __WIN32__) && ! defined __CYGWIN__  
       if (ptid.p != PTHREAD_NA.p) {
 #else
-      if (ptid != PTHREAD_NA) {
+         if ((int64)ptid != PTHREAD_NA) {
 #endif
 	 if (!joined)
 	    pthread_detach(ptid);
